Build fixes - Merged two configs into one. - Made package.json reading fail safe. This also fixes build failing in previous commit.
